# 02-coding-interview/openapi.yaml
openapi: 3.0.3
...
info:
  title: Online Coding Interview Platform API
  description: |
    ## 🎯 Overview
    This API powers an online coding interview platform that enables real-time collaborative coding sessions.
    Interviewers can create sessions, share links with candidates, and watch them code in real-time.
    
    ## ✨ Key Features
    - **Session Management**: Create and join interview sessions with unique IDs
    - **Real-time Collaboration**: WebSocket-based live code sharing
    - **Multi-language Support**: JavaScript, Python, Java, C++ and more
    - **Code Execution**: Safe code execution with output streaming
    - **Simple Authentication**: Session-based access with shareable links
    
    ## 🔌 WebSocket Endpoint
    For real-time collaboration, connect to: `ws://api-server/ws/sessions/{sessionId}`
    
    See the WebSocket Messages section below for message formats.
    
    ## 🚀 Quick Start
    1. Create a session using `POST /api/v1/sessions`
    2. Share the `joinUrl` with participants
    3. Connect to the WebSocket endpoint for real-time updates
    4. Execute code using `POST /api/v1/sessions/{sessionId}/execute`
  version: 1.0.0
  contact:
    name: API Support
    email: support@codinginterview.platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

# SERVER CONFIGURATION
# This tells clients where your API is hosted
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
...

...
paths:
...
  /sessions:
    post:
      tags:
        - Sessions
      summary: Create a new interview session
      description: |
        Creates a new coding interview session and returns a unique session ID.
        The response includes a shareable URL that participants can use to join.
        
        **Use Case**: An interviewer clicks "Create New Interview" to start a session.
      operationId: create_session
      requestBody:
        description: Optional session configuration
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              hostName: "John Interviewer"
              initialLanguage: "javascript"
              sessionName: "Frontend Developer Interview"
              maxParticipants: 5
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
              example:
                sessionId: "abc123xyz789"
                joinUrl: "http://localhost:3000/session/abc123xyz789"
                createdAt: "2024-01-15T10:30:00Z"
                hostName: "John Interviewer"
                status: "active"
                participantCount: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Sessions
      summary: List active sessions
      description: |
        Returns a list of all active sessions. Useful for debugging or admin purposes.
        In production, this might be restricted to authenticated admins only.
      operationId: list_sessions
      parameters:
        - name: status
          in: query
          description: Filter sessions by status
          required: false
          schema:
            type: string
            enum: [active, completed, all]
            default: active
        - name: limit
          in: query
          description: Maximum number of sessions to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionSummary'
                  total:
                    type: integer
                    description: Total number of sessions
              example:
                sessions:
                  - sessionId: "abc123xyz789"
                    createdAt: "2024-01-15T10:30:00Z"
                    hostName: "John Interviewer"
                    participantCount: 2
                    status: "active"
                  - sessionId: "def456uvw012"
                    createdAt: "2024-01-15T09:00:00Z"
                    hostName: "Jane HR"
                    participantCount: 1
                    status: "completed"
                total: 2
  /sessions/{sessionId}:
    get:
      tags:
        - Sessions
      summary: Get session details
      description: |
        Retrieves detailed information about a specific session including
        current code, language, participants, and status.
        
        **Use Case**: When a user joins a session, fetch current state to sync their editor.
      operationId: get_session
      parameters:
        - name: sessionId
          in: path
          description: Unique session identifier
          required: true
          schema:
            type: string
            pattern: '^[a-zA-Z0-9-_]+$'
          example: "abc123xyz789"
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetails'
              example:
                sessionId: "abc123xyz789"
                joinUrl: "http://localhost:3000/session/abc123xyz789"
                createdAt: "2024-01-15T10:30:00Z"
                updatedAt: "2024-01-15T10:45:00Z"
                hostName: "John Interviewer"
                status: "active"
                participantCount: 2
                participants:
                  - clientId: "client-001"
                    displayName: "John Interviewer"
                    role: "host"
                    joinedAt: "2024-01-15T10:30:00Z"
                  - clientId: "client-002"
                    displayName: "Jane Candidate"
                    role: "participant"
                    joinedAt: "2024-01-15T10:32:00Z"
                currentCode: |
                  function fibonacci(n) {
                    if (n <= 1) return n;
                    return fibonacci(n - 1) + fibonacci(n - 2);
                  }
                  console.log(fibonacci(10));
                currentLanguage: "javascript"
                executionHistory:
                  - timestamp: "2024-01-15T10:40:00Z"
                    language: "javascript"
                    output: "55"
                    duration: 12
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Sessions
      summary: End a session
      description: |
        Marks a session as completed and disconnects all participants.
        Only the session host should be able to end the session.
      operationId: end_session
      parameters:
        - name: sessionId
          in: path
          description: Unique session identifier
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Session ended successfully
        '404':
          $ref: '#/components/responses/NotFound'

  # LANGUAGE SUPPORT ENDPOINTS
...
  /languages/{languageId}/template:
    get:
      tags:
        - Languages
      summary: Get language template
      description: |
        Returns the default code template for a specific language.
        Useful when switching languages to provide starter code.
      operationId: get_language_template
      parameters:
        - name: languageId
          in: path
          description: Language identifier (e.g., "javascript", "python")
          required: true
          schema:
            type: string
          example: "javascript"
      responses:
        '200':
          description: Language template
          content:
            application/json:
              schema:
                type: object
                properties:
                  languageId:
                    type: string
                  template:
                    type: string
                    description: Default code template
              example:
                languageId: "javascript"
                template: |
                  function solution() {
                    // Your code here
                    console.log("Hello, World!");
                  }
                  
                  solution();
        '404':
          $ref: '#/components/responses/NotFound'

  # CODE EXECUTION ENDPOINTS
  /sessions/{sessionId}/execute:
    post:
      tags:
        - Execution
      summary: Execute code
      description: |
        Executes code for a specific session and language. The code runs in a
        sandboxed environment with resource limits and timeout protection.
        
        **JavaScript**: Can run in the browser, but backend execution provides consistency
        **Other languages**: Must run on the backend
        
        **Security**: Code runs with limited resources, network access disabled,
        and a maximum execution time of 5 seconds.
      operationId: execute_code
      parameters:
        - name: sessionId
          in: path
          description: Session identifier
          required: true
          schema:
            type: string
          example: "abc123xyz789"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteCodeRequest'
            example:
              code: |
                for i in range(1, 11):
                    print(f"Number: {i}")
              language: "python"
              stdin: ""
              timeLimit: 5000
      responses:
        '200':
          description: Execution completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
              examples:
                success:
                  summary: Successful execution
                  value:
                    sessionId: "abc123xyz789"
                    language: "python"
                    stdout: "Number: 1\nNumber: 2\nNumber: 3\nNumber: 4\nNumber: 5\nNumber: 6\nNumber: 7\nNumber: 8\nNumber: 9\nNumber: 10\n"
                    stderr: ""
                    exitCode: 0
                    duration: 15
                    success: true
                    timestamp: "2024-01-15T10:45:00Z"
                error:
                  summary: Runtime error
                  value:
                    sessionId: "abc123xyz789"
                    language: "python"
                    stdout: ""
                    stderr: "NameError: name 'undefined_variable' is not defined"
                    exitCode: 1
                    duration: 8
                    success: false
                    error: "Runtime error occurred"
                    timestamp: "2024-01-15T10:46:00Z"
                timeout:
                  summary: Execution timeout
                  value:
                    sessionId: "abc123xyz789"
                    language: "python"
                    stdout: "Processing..."
                    stderr: ""
                    exitCode: -1
                    duration: 5000
                    success: false
                    error: "Execution timed out after 5 seconds"
                    timestamp: "2024-01-15T10:47:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

# WEBSOCKET DOCUMENTATION
# Note: OpenAPI doesn't formally support WebSocket, but we document it here
...
components:
...
  schemas:
...
    CreateSessionRequest:
      type: object
      description: Optional configuration when creating a session
      properties:
        hostName:
          type: string
          description: Display name of the session host
          maxLength: 50
          example: "John Interviewer"
        sessionName:
          type: string
          description: Optional session title
          maxLength: 100
          example: "Frontend Developer Interview"
        initialLanguage:
          type: string
          description: Starting programming language
          enum: [javascript, python, java, cpp]
          default: javascript
        maxParticipants:
          type: integer
          description: Maximum number of participants allowed
          minimum: 2
          maximum: 10
          default: 5
    
    # Session Response
    SessionResponse:
      type: object
      description: Created session information
      required:
        - sessionId
        - joinUrl
        - createdAt
        - status
      properties:
        sessionId:
          type: string
          description: Unique session identifier
          example: "abc123xyz789"
        joinUrl:
          type: string
          format: uri
          description: URL to share with participants
          example: "http://localhost:3000/session/abc123xyz789"
        createdAt:
          type: string
          format: date-time
          description: Session creation timestamp
        hostName:
          type: string
          description: Display name of the host
        status:
          type: string
          enum: [active, completed]
          description: Current session status
        participantCount:
          type: integer
          description: Current number of participants
    
    # Session Summary (for list view)
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "VALIDATION_ERROR"
            message: "Invalid request parameters"
            details:
              errors:
                - field: "language"
                  message: "Unsupported language: cobol"
            timestamp: "2024-01-15T10:30:00Z"
...
  securitySchemes:
    sessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
      description: |
        Session token for authentication (future implementation).
        For now, sessions are identified by their ID in the URL.
    
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication for admin endpoints (future implementation).
        Not required for MVP.

# ============================================================================
# HOW THE REACT FRONTEND AND FASTAPI BACKEND USE THIS API
# ============================================================================
# 
# ## React Frontend Integration
# 
# The React frontend (already built) will consume this API as follows:
# 
# 1. **Landing Page**:
#    - Calls `POST /api/v1/sessions` when creating a new interview
#    - Redirects to the session page with the returned `sessionId`
# 
# 2. **Session Page**:
#    - Calls `GET /api/v1/sessions/{sessionId}` to load initial state
#    - Calls `GET /api/v1/languages` to populate language dropdown
#    - Connects to `ws://localhost:8000/ws/sessions/{sessionId}` for real-time updates
#    - Sends WebSocket messages for code changes and receives updates from other users
#    - Calls `POST /api/v1/sessions/{sessionId}/execute` for non-JavaScript languages
# 
# 3. **WebSocket Flow**:
#    - On connect: Send `join_session` message
#    - On code change: Send `code_update` message
#    - On language change: Send `language_change` message
#    - Listen for: `user_joined`, `user_left`, `code_update`, `language_change`, `execution_result`
...